+++ src/main/java/com/yeoro/twogether/domain/waypoint/dto/response/WaypointSummaryResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 public record WaypointSummaryResponse(
     String name,
     String imageUrl,
-    Long order
+    int order
 ) {
 
 }
